test_cluster_single_rack: split into v3 and v4 variants
