+++ shotover-proxy/example-configs/cassandra-cluster/docker-compose-cassandra-v3.yml
