+++ shotover-proxy/example-configs/cassandra-cluster/docker-compose-cassandra-v4.yml
@@ -34,13 +34,6 @@ services:
       HEAP_NEWSIZE: "48M"
       CASSANDRA_ENABLE_SCRIPTED_USER_DEFINED_FUNCTIONS: "true"
       CASSANDRA_ENABLE_USER_DEFINED_FUNCTIONS: "true"
-      CASSANDRA_KEYSTORE_PASSWORD: "password"
-      CASSANDRA_TRUSTSTORE_PASSWORD: "password"
-      CASSANDRA_CLIENT_ENCRYPTION: "true"
-    volumes:
-      &volumes
-      - ../cassandra-tls/certs/keystore.p12:/bitnami/cassandra/secrets/keystore
-      - ../cassandra-tls/certs/truststore.p12:/bitnami/cassandra/secrets/truststore
 
   cassandra-two:
     image: bitnami/cassandra:4.0.6
@@ -49,7 +42,6 @@ services:
         ipv4_address: 172.16.1.3
     healthcheck: *healthcheck
     environment: *environment
-    volumes: *volumes
 
   cassandra-three:
     image: bitnami/cassandra:4.0.6
@@ -58,4 +50,3 @@ services:
         ipv4_address: 172.16.1.4
     healthcheck: *healthcheck
     environment: *environment
-    volumes: *volumes
